requirements.txt
absl-py==2.1.0
json5==0.9.25
keras==3.5.0
keras-nlp==0.15.0
model-alignment==0.1
pandas==2.2.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metal==1.1.0